Reorganize the physics module; update lib.rs docs
- static.rs -> static_mesh.rs; as this module deals with generating
static meshes from particles
- rigid_bodies.rs -> partilce_colliders; as this module deals with
generating avian colliders which particles will recognize and avoid
during movement processing
- Move the majority of geometric algorithms to the utils module. Move a
helper function to static_mesh from physics::geometry
