+++ src/lib.rs
@@ -38,7 +38,7 @@
 //! | [`mod@render`]       | Particle color profiles and chunk-based rendering                        | —                          |
 //! | [`mod@movement`]     | Particle movement systems                                                | —                          |
 //! | [`mod@reactions`]    | Inter-particle reactions                                                 | `render`, `movement`       |
-//! | [`mod@physics`]      | [avian2d](https://docs.rs/avian2d) static and dynamic rigid body integration | —                          |
+//! | [`mod@physics`]      | [avian2d](https://docs.rs/avian2d) bridge between rigid bodies and particles | —                          |
 //! | [`mod@debug`]        | Debug resources                                                          | —                          |
 //! | [`mod@persistence`]  | Save/load chunks, particle types, and scenes to disk                     | `bfs`, `bfc`               |
 //! | `bfs`                | Enables [`persistence::bfs`] — compact particle scene format             | —                          |
@@ -170,12 +170,23 @@
 //!
 //! - [Dynamic rigid bodies](`crate::physics::dynamic`) — promote particles into physics-driven
 //!   rigid body proxies and rejoin them back into the simulation
-//! - [Static rigid bodies](`crate::physics::static`) — per-chunk collision mesh generation from
+//! - [Static rigid bodies](`crate::physics::static_mesh`) — per-chunk collision mesh generation from
 //!   marked particles
+//! - [Particle colliders](`crate::physics::particle_colliders`) — make Avian colliders occupy
+//!   particle-grid cells so falling sand movement routes around them
 //! - [Tagging particles as static rigid bodies](`StaticRigidBodyParticle`)
 //! - [Promoting particles to dynamic rigid bodies](`crate::physics::PromoteDynamicRigidBodyParticle`)
+//! - [Marking Avian colliders as particle blockers](`ParticleCollider`)
+//! - [Configuring collider resting behavior](`ParticleColliderRestingOptions`)
 //! - [Configuring collision mesh calculation intervals](StaticMeshUpdateInterval)
 //! - [Configuring polygon simplification tolerances](`DouglasPeuckerEpsilon`)
+//!
+//! ## [Grid geometry utilities](`crate::utils::geometry`)
+//!
+//! - [Grouping connected cells](`crate::utils::connected_components`)
+//! - [Finding perimeter cells](`crate::utils::perimeter_positions`)
+//! - [Simplifying polylines](`crate::utils::simplify_polyline`)
+//! - [Generating meshes from grid cells](`crate::utils::mesh_from_grid_cells`)
 //!
 //! ## [World persistence](crate::persistence)
 //!
