+++ src/physics/mod.rs
@@ -1,6 +1,7 @@
 //! Integrates [avian2d](https://docs.rs/avian2d) physics with the falling sand simulation.
 //!
-//! This module provides two sub-modules, each owning a distinct domain:
+//! Physics bridges two worlds: the discrete particle grid and continuous rigid body simulation.
+//! This module provides three integration paths:
 //!
 //! - **[`dynamic`]** — sending a [`PromoteDynamicRigidBodyParticle`] for a particle removes it from the
 //!   simulation and spawns a separate physics-driven rigid body entity at its position. Each
@@ -8,13 +9,16 @@
 //!   [`ParticleMap`](crate::ParticleMap) (or at the map edge), the rigid body is despawned and
 //!   the original particle is restored to the simulation at the nearest vacant position.
 //!
-//! - **[`static`](mod@self::static)** — particles marked with [`StaticRigidBodyParticle`] (on the
-//!   [`ParticleType`](crate::ParticleType)) contribute to per-chunk collision meshes.
+//! - **[`static_mesh`]** — particles marked with [`StaticRigidBodyParticle`] (on the
+//!   [`ParticleType`](crate::ParticleType)) contribute to per-chunk static collision meshes.
+//!
+//! - **[`particle_colliders`]** — regular Avian colliders marked with [`ParticleCollider`] occupy
+//!   particle-grid cells through [`RigidBodyParticleOccupancy`], allowing falling sand movement to
+//!   route around rigid bodies.
 //!
 //! ## Static mesh generation
 //!
-//! The static mesh generation pipeline runs each frame and works
-//! per-chunk:
+//! The static mesh generation pipeline runs each frame and works per-chunk:
 //!
 //! 1. **Identify dirty chunks** — chunks whose dirty state just cleared ("settled") are
 //!    processed immediately. Chunks still actively dirty are throttled by
@@ -24,14 +28,14 @@
 //!    record which cells contain a `StaticRigidBodyParticle` entity. Compare against
 //!    the cached bitmap from the previous pass; skip if unchanged.
 //!
-//! 3. **Spawn async mesh generation tasks** — for changed chunks, an async task performs:
+//! 3. **Spawn async mesh generation tasks** — for changed chunks, an async task uses
+//!    [`utils::geometry`](crate::utils::geometry) to perform:
 //!    - **Flood-fill** to discover connected components of occupied cells.
 //!    - **Perimeter extraction** — for each component, find the boundary edges between
 //!      occupied and empty cells.
 //!    - **Edge ordering** — assemble edges into closed loops.
 //!    - **Douglas-Peucker simplification** — reduce vertex count using the epsilon from
-//!      [`DouglasPeuckerEpsilon`]. TODO: This needs more in-depth profiling to see if
-//!      we actually get gains from this
+//!      [`DouglasPeuckerEpsilon`].
 //!    - **Ear-cut triangulation** — convert the simplified polygon into a triangle mesh.
 //!
 //! 4. **Poll completed tasks** — merge the per-component meshes into a single trimesh
@@ -43,21 +47,21 @@
 //!    respond to the new collision geometry.
 
 pub mod dynamic;
-pub(crate) mod geometry;
-pub mod rigid_bodies;
-pub mod r#static;
+pub mod particle_colliders;
+pub mod static_mesh;
 
 use avian2d::prelude::PhysicsInterpolationPlugin;
 use bevy::prelude::*;
 
 pub use dynamic::{DynamicRigidBodyProxy, PromoteDynamicRigidBodyParticle, SuspendedParticle};
-pub use rigid_bodies::*;
-pub use r#static::{DouglasPeuckerEpsilon, StaticMeshUpdateInterval};
+pub use particle_colliders::*;
+pub use static_mesh::{DouglasPeuckerEpsilon, StaticMeshUpdateInterval};
 
 use dynamic::DynamicPlugin;
 use dynamic::{promote_dynamic_rigid_bodies, rejoin_dynamic_rigid_bodies};
-use r#static::{StaticPlugin, calculate_static_rigid_bodies};
-pub use r#static::{StaticRigidBodyParticle, StaticRigidBodyParticleCollider};
+use particle_colliders::RigidBodiesPlugin;
+use static_mesh::{StaticPlugin, calculate_static_rigid_bodies};
+pub use static_mesh::{StaticRigidBodyParticle, StaticRigidBodyParticleCollider};
 
 use crate::movement::ParticleMovementSystems;
 use crate::physics::dynamic::sync_dynamic_rigid_bodies_with_particles;
