+++ src/utils/mod.rs
@@ -1,4 +1,4 @@
-//! Common utiltiies for navigating particles
+//! Common utilities for working with particle-grid data.
 
 pub mod geometry;
